Make traits compatible with WASM (#107)
* web-transport-trait: Introduce MaybeSend/MaybeSync

These new types are equivalent to Send and Sync only on non-WASM
platforms.

On WASM, the types exposed by web_sys cannot be used in a multi-threaded
environment, so they are neither Send or Sync. However, we still want
to use a common trait across all platforms.

* web-transport-wasm: Replace &mut self with &self

The Session methods do not require a mutable reference to self.

This is consistant with the web-transport-trait::Session trait.

* web-transport-wasm: Only compile for WASM

Fail with an error if web-transport-wasm is compiled for a non-WASM
target.

* web-transport-wasm: Implement common traits

Since the web-transport-trait traits do not require Send or Sync on
WASM, they can now be implemented for the web-transport-wasm types.

* Fix the build.

---------

Co-authored-by: Luke Curley <kixelated@gmail.com>